Breaking up the sandbox module
