+++ sandbox/src-js/example/Counter.scala
@@ -1,0 +1,25 @@
+package example
+
+import tyrian.*
+import tyrian.Html.*
+
+object Counter:
+
+  opaque type Model = Int
+
+  def init: Model = 0
+
+  enum Msg:
+    case Increment, Decrement
+
+  def view(model: Model): Html[Msg] =
+    div(
+      button(onClick(Msg.Decrement))(text("-")),
+      div(text(model.toString)),
+      button(onClick(Msg.Increment))(text("+"))
+    )
+
+  def update(msg: Msg, model: Model): Model =
+    msg match
+      case Msg.Increment => model + 1
+      case Msg.Decrement => model - 1
