+++ sandbox/src-js/example/Fruit.scala
@@ -1,0 +1,3 @@
+package example
+
+final case class Fruit(name: String, available: Boolean)
