+++ sandbox/src-js/example/Vector2.scala
@@ -1,0 +1,3 @@
+package example
+
+final case class Vector2(x: Double, y: Double)
